Fix model structure and optional variable
