--- week4/Calculator/Calculator/Model/ArithmeticOperation.swift
@@ -1,15 +1,0 @@
-//
-//  ArithmeticOperation.swift
-//  Calculator
-//
-//  Created by 임주민 on 2022/08/21.
-//
-
-import Foundation
-
-enum InputState {
-  case newPreviousNumber
-  case newNextNumber
-  case ongoingPreviousNumber
-  case ongoingNextNumber
-}
