--- week4/Calculator/Calculator/Model/InputState.swift
@@ -1,15 +1,0 @@
-//
-//  InputState.swift
-//  Calculator
-//
-//  Created by 임주민 on 2022/08/21.
-//
-
-import Foundation
-
-enum ArithmeticOperation {
-  case multiply
-  case divide
-  case minus
-  case plus
-}
